Have sel4-build.git go to .build/.
When it was targeting tools/ "repo status" reports that the directories
under tools/ were not under repo control. Using a  hidden directory like
.build/ actually seems better.

There is still a problem, we should be able to add an app to apps/ or
libraries to libs/ and be able to do execute "make config" or
"make menuconfig" and then do a "make" without modifying the build system.